+++ libs/cardano-ledger-test/src/Test/Cardano/Ledger/Generic/PrettyCore.hs
@@ -73,6 +73,7 @@ import Cardano.Ledger.BaseTypes (
   EpochInterval (..),
   EpochNo (..),
   FixedPoint,
+  Mismatch (..),
   Network (..),
   Nonce (..),
   ProtVer (..),
@@ -219,7 +220,6 @@ import Cardano.Ledger.Shelley.Rules (
   LedgerEnv (..),
   PoolEnv (..),
   ShelleyLedgerPredFailure (..),
-  UpecPredFailure,
   UtxoEnv (..),
  )
 import Cardano.Ledger.Shelley.Rules as Shelley (
@@ -231,13 +231,11 @@ import Cardano.Ledger.Shelley.Rules as Shelley (
   ShelleyEpochPredFailure (..),
   ShelleyLedgersPredFailure (..),
   ShelleyNewEpochPredFailure (..),
-  ShelleyNewppPredFailure (..),
   ShelleyPoolPredFailure (..),
   ShelleyPoolreapPredFailure,
   ShelleyPpupPredFailure (..),
   ShelleySnapPredFailure,
   ShelleyTickPredFailure (..),
-  ShelleyUpecPredFailure (..),
   ShelleyUtxoPredFailure (..),
   ShelleyUtxowPredFailure (..),
  )
@@ -1304,18 +1302,6 @@ ppEPOCH Alonzo x = ppShelleyEpochPredFailure x
 ppEPOCH Babbage x = ppShelleyEpochPredFailure x
 ppEPOCH Conway _ = ppString "PredicateFailure (ConwayEPOCH era) = Void, and can never Fail"
 
--- | A bit different since it is NOT of the form: PredicateFailure (EraRule "UPEC" era)
---   but instead, the type family UpecPredFailurePV pv era
---   where, type UpecPredFailure era = UpecPredFailurePV (ProtVerLow era) era
---   But the effect is still the same. The Proof era, fixes the type family result.
-ppUPEC :: Proof era -> UpecPredFailure era -> PDoc
-ppUPEC Shelley x = ppUpecPredicateFailure x
-ppUPEC Mary x = ppUpecPredicateFailure x
-ppUPEC Alonzo x = ppUpecPredicateFailure x
-ppUPEC Allegra x = ppUpecPredicateFailure x
-ppUPEC Babbage x = ppUpecPredicateFailure x
-ppUPEC Conway x = absurd x
-
 -- | A bit different since it is NOT of the form: PredicateFailure (EraRule "LEDGERS" era)
 --   but instead:  State (EraRule "LEDGERS" era)
 --   But the effect is still the same. The Proof era, fixes the type family result.
@@ -1642,17 +1628,17 @@ instance
 -- ================
 
 ppBbodyPredicateFailure :: forall era. Reflect era => ShelleyBbodyPredFailure era -> PDoc
-ppBbodyPredicateFailure (WrongBlockBodySizeBBODY x y) =
+ppBbodyPredicateFailure (WrongBlockBodySizeBBODY (Mismatch supplied expected)) =
   ppRecord
     "WrongBlockBodySizeBBODY"
-    [ ("actual computed BBody size", ppInt x)
-    , ("claimed BBody Size in Header", ppInt y)
+    [ ("actual computed BBody size", ppInt supplied)
+    , ("claimed BBody Size in Header", ppInt expected)
     ]
-ppBbodyPredicateFailure (InvalidBodyHashBBODY h1 h2) =
+ppBbodyPredicateFailure (InvalidBodyHashBBODY (Mismatch supplied expected)) =
   ppRecord
     "(InvalidBodyHashBBODY"
-    [ ("actual hash", ppHash h1)
-    , ("claimed hash", ppHash h2)
+    [ ("actual hash", ppHash supplied)
+    , ("claimed hash", ppHash expected)
     ]
 ppBbodyPredicateFailure (LedgersFailure x) =
   ppSexp "LedgersFailure" [ppLEDGERS @era reify x]
@@ -1738,12 +1724,13 @@ instance PrettyA (ConwayNewEpochPredFailure era) where
 
 -- ===============
 
-ppShelleyEpochPredFailure :: forall era. Reflect era => ShelleyEpochPredFailure era -> PDoc
+ppShelleyEpochPredFailure :: forall era. ShelleyEpochPredFailure era -> PDoc
 ppShelleyEpochPredFailure (PoolReapFailure _) =
   ppString "PoolreapPredicateFailure has no constructors"
 ppShelleyEpochPredFailure (SnapFailure _) =
   ppString "SnapPredicateFailure has no constructors"
-ppShelleyEpochPredFailure (UpecFailure x) = ppUPEC @era reify x
+ppShelleyEpochPredFailure (UpecFailure _) =
+  ppString "UpecPredicateFailure has no constructors"
 
 instance Reflect era => PrettyA (ShelleyEpochPredFailure era) where
   prettyA = ppShelleyEpochPredFailure
@@ -1755,25 +1742,6 @@ instance PrettyA (ShelleyPoolreapPredFailure era) where
 -- This type has no constructors, so the show instance is fine.
 instance PrettyA (ShelleySnapPredFailure era) where
   prettyA = viaShow
-
--- ===============
-
-ppUpecPredicateFailure :: ShelleyUpecPredFailure era -> PDoc
-ppUpecPredicateFailure (NewPpFailure x) = ppNewppPredicateFailure x
-
-instance PrettyA (ShelleyUpecPredFailure era) where
-  prettyA = ppUpecPredicateFailure
-
--- ===============
-ppNewppPredicateFailure :: ShelleyNewppPredFailure era -> PDoc
-ppNewppPredicateFailure (UnexpectedDepositPot c1 c2) =
-  ppRecord
-    "UnexpectedDepositPot"
-    [ ("The total outstanding deposits", pcCoin c1)
-    , ("The deposit pot", pcCoin c2)
-    ]
-
-instance PrettyA (ShelleyNewppPredFailure era) where prettyA = ppNewppPredicateFailure
 
 -- =========================================
 -- Predicate Failure for Alonzo UTXOW
@@ -1835,10 +1803,12 @@ ppShelleyUtxowPredFailure (MissingTxBodyMetadataHash m) =
   ppSexp " MissingTxMetadata" [ppAuxiliaryDataHash m]
 ppShelleyUtxowPredFailure (MissingTxMetadata m) =
   ppSexp " MissingTxMetadata" [ppAuxiliaryDataHash m]
-ppShelleyUtxowPredFailure (ConflictingMetadataHash h1 h2) =
+ppShelleyUtxowPredFailure (ConflictingMetadataHash (Mismatch supplied expected)) =
   ppRecord
     "ConflictingMetadataHash"
-    [("Hash in the body", ppAuxiliaryDataHash h1), ("Hash of full metadata", ppAuxiliaryDataHash h2)]
+    [ ("Hash in the body", ppAuxiliaryDataHash supplied)
+    , ("Hash of full metadata", ppAuxiliaryDataHash expected)
+    ]
 ppShelleyUtxowPredFailure InvalidMetadata =
   ppSexp "InvalidMetadata" []
 ppShelleyUtxowPredFailure (ExtraneousScriptWitnessesUTXOW m) =
@@ -1995,9 +1965,8 @@ ppShelleyPoolPredFailure (StakePoolNotRegisteredOnKeyPOOL kh) =
     ]
 ppShelleyPoolPredFailure
   ( StakePoolRetirementWrongEpochPOOL
-      curEpoch
-      poolRetEpoch
-      firstTooFarEpoch
+      (Mismatch {mismatchExpected = curEpoch})
+      (Mismatch {mismatchSupplied = poolRetEpoch, mismatchExpected = firstTooFarEpoch})
     ) =
     ppRecord
       "StakePoolRetirementWrongEpochPOOL"
@@ -2007,8 +1976,7 @@ ppShelleyPoolPredFailure
       ]
 ppShelleyPoolPredFailure
   ( StakePoolCostTooLowPOOL
-      prcStakePoolCost
-      ppStakePoolCost
+      (Mismatch {mismatchSupplied = prcStakePoolCost, mismatchExpected = ppStakePoolCost})
     ) =
     ppRecord
       "StakePoolCostTooLowPOOL"
@@ -2017,8 +1985,7 @@ ppShelleyPoolPredFailure
       ]
 ppShelleyPoolPredFailure
   ( WrongNetworkPOOL
-      nwId
-      regCertNwId
+      (Mismatch {mismatchSupplied = nwId, mismatchExpected = regCertNwId})
       stakePoolId
     ) =
     ppRecord
@@ -2153,7 +2120,7 @@ ppShelleyUtxoPredFailure (Shelley.BadInputsUTxO x) =
   ppSexp "BadInputsUTxO" [ppSet pcTxIn x]
 ppShelleyUtxoPredFailure (Shelley.ExpiredUTxO ttl slot) =
   ppRecord "ExpiredUTxO" [("transaction time to live", pcSlotNo ttl), ("current slot", pcSlotNo slot)]
-ppShelleyUtxoPredFailure (Shelley.MaxTxSizeUTxO actual maxs) =
+ppShelleyUtxoPredFailure (Shelley.MaxTxSizeUTxO (Mismatch {mismatchSupplied = actual, mismatchExpected = maxs})) =
   ppRecord
     "MaxTxSizeUTxO"
     [ ("Actual", ppInteger actual)
@@ -2161,7 +2128,7 @@ ppShelleyUtxoPredFailure (Shelley.MaxTxSizeUTxO actual maxs) =
     ]
 ppShelleyUtxoPredFailure (Shelley.InputSetEmptyUTxO) =
   ppSexp "InputSetEmptyUTxO" []
-ppShelleyUtxoPredFailure (Shelley.FeeTooSmallUTxO computed supplied) =
+ppShelleyUtxoPredFailure (Shelley.FeeTooSmallUTxO (Mismatch {mismatchSupplied = supplied, mismatchExpected = computed})) =
   ppRecord
     "FeeTooSmallUTxO"
     [ ("min fee for this transaction", pcCoin computed)
@@ -2203,7 +2170,7 @@ instance Reflect era => PrettyA (ShelleyUtxoPredFailure era) where
 -- Predicate Failure for Shelley PPUP
 
 ppPpupPredicateFailure :: ShelleyPpupPredFailure era -> PDoc
-ppPpupPredicateFailure (NonGenesisUpdatePPUP x y) =
+ppPpupPredicateFailure (NonGenesisUpdatePPUP (Mismatch {mismatchSupplied = x, mismatchExpected = y})) =
   ppRecord
     "NonGenesisUpdatePPUP"
     [ ("KeyHashes which are voting", ppSet pcKeyHash x)